requirements.txt
streamlit
opencv-python
face_recognition
numpy
Pillow
sentence-transformers
transformers
faiss-cpu
